// src/app/core/models/ICart/icart.interface.ts
export interface ICart {
    status: string;
    numOfCartItems: number;
    cartId: string;
    data: Data;
}

interface Data {
    _id: string;
    cartOwner: string;
    products: Product2[];
    createdAt: string;
    updatedAt: string;
    __v: number;
    totalCartPrice: number;
}

interface Product2 {
    count: number;
    _id: string;
    product: Product;
    price: number;
}

interface Product {
    subcategory: Subcategory[];
    _id: string;
    title: string;
    quantity: number;
    imageCover: string;
    category: Category;
    brand: Category;
    ratingsAverage: number;
    id: string;
}

interface Category {
    _id: string;
    name: string;
    slug: string;
    image: string;
}

interface Subcategory {
    _id: string;
    name: string;
    slug: string;
    category: string;
}